core/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-memory-core"
version = "1.2.0"
description = "Core logic for autonomous agent memory management"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0.0,<3.0.0",
    "pyyaml>=6.0,<7.0.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["adapters*", "api*", "config*", "core*", "reasoning*", "stores*"]
